lib: picker: SeslColorSwatchView: fix spacing between colors

+++ picker/picker-color/src/main/java/androidx/picker/widget/SeslColorSwatchView.java
@@ -151,10 +151,10 @@ class SeslColorSwatchView extends View {
         for (int i = 0; i < SWATCH_ITEM_COLUMN; i++) {
             for (int j = 0; j < SWATCH_ITEM_ROW; j++) {
                 paint.setColor(mColorSwatch[i][j]);
-                canvas.drawRect(i * mSwatchItemWidth + SWATCH_ITEM_SIZE_ROUNDING_VALUE,
-                        j * mSwatchItemHeight + SWATCH_ITEM_SIZE_ROUNDING_VALUE,
-                        mSwatchItemWidth * (i + 1) + SWATCH_ITEM_SIZE_ROUNDING_VALUE,
-                        mSwatchItemHeight * (j + 1) + SWATCH_ITEM_SIZE_ROUNDING_VALUE,
+                canvas.drawRect((int) (i * mSwatchItemWidth + SWATCH_ITEM_SIZE_ROUNDING_VALUE),
+                        (int) (j * mSwatchItemHeight + SWATCH_ITEM_SIZE_ROUNDING_VALUE),
+                        (int) (mSwatchItemWidth * (i + 1) + SWATCH_ITEM_SIZE_ROUNDING_VALUE),
+                        (int) (mSwatchItemHeight * (j + 1) + SWATCH_ITEM_SIZE_ROUNDING_VALUE),
                         paint);
             }
         }
